Monitor parent process to ensure cleanup
In the event that the creator of a stats_hero process dies before cleaning up
the stats_hero worker, that process can be leaked.

This patch monitors the parent process.  If the parent process goes down prior
to cleaning up the stats_hero worker, the stats_hero worker will log a message
and terminate cleanly.